+++ src/ejercicio_4/vite-project/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Vite App</title>
+  </head>
+  <body>
+    <button id="main" type="button">Iniciar Debug</button>
+    <script type="module" src="/main.js"></script>
+  </body>
+</html>
